% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{github_issues}
\alias{github_issues}
\alias{github_issues_training}
\alias{github_issues_eval}
\title{Documentation Github Issues}
\format{A data frame with 3 variables:
\describe{
\item{text}{Text of the issue}
\item{label}{Type of issue. "DOCUMENTATION" for all samples}
\item{answer}{"accept" if the label applies, "reject" if it doesn't apply.}
...
}}
\source{
\url{https://github.com/explosion/projects/tree/master/textcat-docs-issues}
}
\usage{
github_issues_training

github_issues_eval
}
\description{
Two datasets, training and evaluation, of labeled Github issues.
}
\keyword{datasets}
